refactor(config-ui): adjust the plugin config
