+++ config-ui/src/plugins/register/bamboo/config.tsx
@@ -17,14 +17,15 @@
  */
 
 import { DOC_URL } from '@/release';
+import { IPluginConfig } from '@/types';
 
-import Icon from './assets/icon.svg';
+import Icon from './assets/icon.svg?react';
 
-export const BambooConfig = {
+export const BambooConfig: IPluginConfig = {
   plugin: 'bamboo',
   name: 'Bamboo',
-  icon: Icon,
-  sort: 11,
+  icon: ({ color }) => <Icon fill={color} />,
+  sort: 2,
   connection: {
     docLink: DOC_URL.PLUGIN.BAMBOO.BASIS,
     fields: [
