+++ config-ui/src/plugins/register/bitbucket/config.tsx
@@ -17,14 +17,15 @@
  */
 
 import { DOC_URL } from '@/release';
+import { IPluginConfig } from '@/types';
 
-import Icon from './assets/icon.svg';
+import Icon from './assets/icon.svg?react';
 
-export const BitBucketConfig = {
+export const BitBucketConfig: IPluginConfig = {
   plugin: 'bitbucket',
   name: 'BitBucket',
-  icon: Icon,
-  sort: 5,
+  icon: ({ color }) => <Icon fill={color} />,
+  sort: 3,
   connection: {
     docLink: DOC_URL.PLUGIN.BITBUCKET.BASIS,
     initialValues: {
