+++ config-ui/src/plugins/register/github/config.tsx
@@ -19,15 +19,16 @@
 import { pick } from 'lodash';
 
 import { DOC_URL } from '@/release';
+import { IPluginConfig } from '@/types';
 
-import Icon from './assets/icon.svg';
+import Icon from './assets/icon.svg?react';
 import { Token, Graphql, GithubApp, Authentication } from './connection-fields';
 
-export const GitHubConfig = {
+export const GitHubConfig: IPluginConfig = {
   plugin: 'github',
   name: 'GitHub',
-  icon: Icon,
-  sort: 1,
+  icon: ({ color }) => <Icon fill={color} />,
+  sort: 4,
   connection: {
     docLink: DOC_URL.PLUGIN.GITHUB.BASIS,
     initialValues: {
