+++ config-ui/src/plugins/register/jenkins/config.tsx
@@ -17,35 +17,48 @@
  */
 
 import { DOC_URL } from '@/release';
+import { IPluginConfig } from '@/types';
 
-import Icon from './assets/icon.svg';
+import Icon from './assets/icon.svg?react';
 
-export const SonarQubeConfig = {
-  plugin: 'sonarqube',
-  name: 'SonarQube',
-  icon: Icon,
-  sort: 7,
+export const JenkinsConfig: IPluginConfig = {
+  plugin: 'jenkins',
+  name: 'Jenkins',
+  icon: ({ color }) => <Icon fill={color} />,
+  sort: 6,
   connection: {
-    docLink: DOC_URL.PLUGIN.SONARQUBE.BASIS,
+    docLink: DOC_URL.PLUGIN.JENKINS.BASIS,
     fields: [
       'name',
       {
         key: 'endpoint',
-        subLabel: 'Provide the SonarQube instance API endpoint. E.g. http://<host>:<port>/api/',
+        subLabel: 'Provide the Jenkins instance API endpoint. E.g. https://api.jenkins.io',
       },
-      'token',
+      'username',
+      'password',
       'proxy',
       {
         key: 'rateLimitPerHour',
         subLabel:
-          'By default, DevLake uses 10,000 requests/hour for data collection for SonarQube. But you can adjust the collection speed by setting up your desirable rate limit.',
-        learnMore: DOC_URL.PLUGIN.SONARQUBE.RATE_LIMIT,
-        externalInfo: 'SonarQube does not specify a maximum value of rate limit.',
+          'By default, DevLake uses 10,000 requests/hour for data collection for Jenkins. But you can adjust the collection speed by setting up your desirable rate limit.',
+        learnMore: DOC_URL.PLUGIN.JENKINS.RATE_LIMIT,
+        externalInfo: 'Jenkins does not specify a maximum value of rate limit.',
         defaultValue: 10000,
       },
     ],
   },
   dataScope: {
-    title: 'Repositories',
+    localSearch: true,
+    title: 'Jobs',
+    millerColumn: {
+      columnCount: 2.5,
+    },
+  },
+  scopeConfig: {
+    entities: ['CICD'],
+    transformation: {
+      deploymentPattern: '(deploy|push-image)',
+      productionPattern: 'prod(.*)',
+    },
   },
 };
