+++ config-ui/src/plugins/register/jira/config.tsx
@@ -17,15 +17,16 @@
  */
 
 import { DOC_URL } from '@/release';
+import { IPluginConfig } from '@/types';
 
-import Icon from './assets/icon.svg';
+import Icon from './assets/icon.svg?react';
 import { Auth } from './connection-fields';
 
-export const JiraConfig = {
+export const JiraConfig: IPluginConfig = {
   plugin: 'jira',
   name: 'Jira',
-  icon: Icon,
-  sort: 3,
+  icon: ({ color }) => <Icon fill={color} />,
+  sort: 7,
   connection: {
     docLink: DOC_URL.PLUGIN.JIRA.BASIS,
     fields: [
