+++ config-ui/src/plugins/register/pagerduty/config.tsx
@@ -17,13 +17,14 @@
  */
 
 import { DOC_URL } from '@/release';
+import { IPluginConfig } from '@/types';
 
-import Icon from './assets/icon.svg';
+import Icon from './assets/icon.svg?react';
 
-export const PagerDutyConfig = {
+export const PagerDutyConfig: IPluginConfig = {
   plugin: 'pagerduty',
   name: 'PagerDuty',
-  icon: Icon,
+  icon: ({ color }) => <Icon fill={color} />,
   sort: 8,
   connection: {
     docLink: DOC_URL.PLUGIN.PAGERDUTY.BASIS,
